The following changes were effected:
(1)   In the presence of N Forbidden lines, TCGenerator now forms
invalid files by deleting N-1 of the Forbidden lines and leaving 1 of
them.
(2)   A new section has been included in the log files, called "Action
Taken on the IDF". The aim of this section is to explain what was
removed from the original IDF to form the new IDF.  This will help in
tracing (quickly) what led to a validation error(s).